sis_byte/backend_byte/requirements.txt
alembic==1.15.2
apispec==6.3.0
click==8.2.0
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
flask-smorest==0.41.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mysql-connector-python==8.0.33
packaging==25.0
protobuf==3.20.3
PyJWT==2.10.1
python-dotenv==1.0.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
webargs==8.7.0
Werkzeug==2.2.2
requests
